# app/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExpenseWise REST API",
    "version": "1.0.0",
    "description": "Developer API for ExpenseWise personal finance tracker. All endpoints return JSON and require authorization headers for private data."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Token",
        "description": "Enter your personal secure API Token generated from settings."
      }
    }
  },
  "paths": {
    "/v1/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user account",
        "description": "Creates a temporary account and triggers an email verification code (OTP) sent to the specified address.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created. OTP code sent successfully.",
            "content": {
              "application/json": {
                "example": { "message": "OTP code sent to email." }
              }
            }
          },
          "400": { "description": "Invalid payload or validation failed." }
        }
      }
    },
    "/v1/auth/verify-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Verify OTP registration code",
        "description": "Verifies the email verification OTP to fully activate the newly registered user account.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "otp": { "type": "string" }
                },
                "required": ["email", "otp"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified and account activated.",
            "content": {
              "application/json": {
                "example": { "message": "Email verified successfully." }
              }
            }
          },
          "400": { "description": "Incorrect or expired verification code." }
        }
      }
    },
    "/v1/auth/resend-otp": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Request a fresh verification OTP",
        "description": "Generates and sends a new registration OTP code via email for an unverified account. Rate-limited to one request per 60 seconds.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "user_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP resent successfully.",
            "content": {
              "application/json": {
                "example": { "message": "A fresh verification OTP has been sent to your email address." }
              }
            }
          },
          "400": { "description": "Account is already verified or not found." },
          "429": { "description": "Too many requests. Please wait before requesting another OTP." }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login and retrieve API token",
        "description": "Exchanges credentials for a valid API Bearer access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful.",
            "content": {
              "application/json": {
                "example": {
                  "token": "TBY4XiyEpfa5b66d4...",
                  "user": {
                    "id": "abc-123",
                    "name": "Indrajit Ghosh",
                    "email": "indrajitghosh912@gmail.com"
                  }
                }
              }
            }
          },
          "401": { "description": "Incorrect credentials." }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Revoke API access token",
        "description": "Deletes the token record on the server database to render it permanently invalid.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Token revoked successfully.",
            "content": {
              "application/json": {
                "example": { "message": "Logged out successfully. Token invalidated." }
              }
            }
          },
          "401": { "description": "Unauthorized. Missing or invalid Bearer token." }
        }
      }
    },
...
  }
}